Fix sliders
