+++ src/index.js
@@ -18,7 +18,6 @@ import './js/layout/сatalog/gallery.js';
 //product
 import './js/layout/product/infoAboutProduct.js';
 // import './js/layout/product/productCreateInfo.js';
-import './js/layout/product/productSlider.js';
 import './js/layout/product/productBin.js';
 import './js/layout/product/recomendationsCategory.js';
 import './js/layout/product/productModalAddToCart.js';
