+++ src/js/layout/fitting/videoSlider.js
@@ -6,61 +6,6 @@ import set from '../../../../src/js/json/videoSet.json';
 export const videoSlider_videoSliderCreate = videoSlider(set);
 // export default videoSlider_videoSliderCreate;
 
-export function openVideoSlider() {
-  //SLIDER SETTINGS
-  $(document).ready(function () {
-    $('.slider').slick({
-      infinite: true,
-      centerMode: true,
-      slidesToShow: 1,
-      slidesToScroll: 1,
-      arrows: false,
-      swipe: true,
-      mobileFirst: true,
-      dots: true,
-      lazyLoad: 'progressive',
-      variableWidth: true,
-      responsive: [
-        {
-          breakpoint: 1377,
-          settings: {
-            dots: true,
-            slidesToShow: 3,
-            slidesToScroll: 3,
-            variableWidth: true,
-            autoPlaySpeed: 2000,
-            pauseFocus: true,
-            pauseHover: true,
-            pauseOnDotsHover: true,
-            swipe: true,
-          },
-        },
-      ],
-      //   dots: true,
-      //   slidesToShow: 3,
-      //   slidesToScroll: 3,
-      //   variableWidth: true,
-      //   autoPlaySpeed: 2000,
-      //   pauseFocus: true,
-      //   pauseHover: true,
-      //   pauseOnDotsHover: true,
-      //   swipe: true,
-
-      //   //  mobileFirst: true,
-      //   responsive: [
-      //     {
-      //       breakpoint: 320,
-      //       settings: {
-      //         slidesToShow: 1,
-      //         slidesToScroll: 1,
-      //         swipe: true,
-      //       },
-      //     },
-      //   ],
-    });
-  });
-}
-
 export function fittingVideoSliderPlayer() {
   const fittingRefs = {
     play: document.querySelectorAll('.video-fitting-svg'),
