+++ src/js/layout/home/hero.js
@@ -4,46 +4,3 @@ import pageHeroMarkupTemplate from '../../../views/partials/home/hero.hbs';
 // Create Markup
 export const pageHeroSliderMarkup = pageHeroMarkupTemplate(pageHeroSliderData);
 
-export function heroSlider() {
-  window.jQuery = window.$ = require('jquery');
-  require('../../slider/slick.min.js');
-  // Slider options
-  $(document).ready(function () {
-    $('.hero__slider').slick({
-      arrows: true,
-      dots: true,
-      adaptiveHeight: true,
-      slidesToShow: 1,
-      slidesToScroll: 1,
-      speed: 1000,
-      easing: 'ease',
-      infinite: true,
-      initialSlide: 0,
-      autoplay: true,
-      autoplaySpeed: 2000,
-      pauseOnFocus: true,
-      pauseOnHover: true,
-      pauseOnDotsHover: true,
-      draggable: true,
-      swipe: true,
-      touchThreshold: 4,
-      touchMove: true,
-      waitForAnimate: true,
-
-      responsive: [
-        {
-          breakpoint: 1377,
-          settings: {
-            arrows: false,
-          },
-        },
-        {
-          breakpoint: 1378,
-          settings: {
-            mobileFirst: true,
-          },
-        },
-      ],
-    });
-  });
-}
