+++ src/sass/layout/fitting/_videoSlider.scss
@@ -1,68 +1,80 @@
 .fitting__first-video,
 .fitting__video {
-  margin-right: 15px;
+  // width: 280px;
   @media screen and (min-width: $desktop-up) {
     width: 185px;
     height: auto;
   }
-  @media screen and (min-width: $mobile-first) and (max-width: $tablet-general) {
+  @media screen and (min-width: $mobile-general) and (max-width: $tablet-general) {
     width: 374px;
     height: auto;
   }
 }
 .fitting__video-slider {
   margin-top: 50px;
-  padding-bottom: 10px;
   margin-bottom: 50px;
-  & .slick-dots li + li{
-    margin-left: 20px;
-  }
-}
-
-//SLICK SLIDER STYLES
-
-/* video slider */
-.slick-list {
-  overflow: hidden;
-}
-.slick-track {
-  display: flex;
-}
-.splide__track {
-  @media screen and (min-width: $desktop-up){
-   max-width: 590px; 
-  }
-  
-}
-
-/* extra */
-
-.slider .slick-arrow {
-  position: absolute;
-
-  z-index: 10;
-  font-size: 0;
-  @media screen and (min-width: $mobile-first) and (max-width: $tablet-general) {
-    display: none;
-  }
-}
-.slider .slick-arrow.slick-prev {
-  left: 460px;
-  top: -40px;
-  background-image: url('../images/arrow-left.svg');
-  width: 40px;
-  height: 15px;
-  border: none;
-  background-color: #f9f9f9;
-  background-repeat: no-repeat;
-}
-.fitting__video-slider {
   position: relative;
   @media screen and (min-width: $mobile-first) and (max-width: $tablet-general) {
     margin-left: auto;
     margin-right: auto;
   }
+  @media screen and (min-width: $desktop-up) {
+    padding: 35px;
+    background-color: $third-background-color;
+  }
+  &-container {
+    overflow: hidden;
+  }
+  & .slider__item {
+    margin-right: 10px;
+  }
+  & .slick-dots li + li {
+    margin-left: 20px;
+  }
+  & .slick-track {
+    display: flex;
+  }
+  & .slick-list {
+    overflow: hidden;
+  }
+  & .splide__track {
+    @media screen and (min-width: $desktop-up) {
+      max-width: 590px;
+    }
+  }
+
+  & .slider .slick-dots {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  & .slider .slick-dots li {
+    list-style: none;
+  }
+  & .slider .slick-dots button {
+    font-size: 0;
+    width: 5px;
+    height: 5px;
+    background-color: $in-instagram-slider-dots;
+    border-radius: 70%;
+    border: none;
+  }
+  & .slider .slick-dots li.slick-active button {
+    width: 26px;
+    border-radius: 20%;
+    @media screen and (min-width: $desktop-up) {
+      width: 66px;
+    }
+  }
+  & .slider__item {
+    position: relative;
+
+    @media screen and (min-width: $mobile-first) {
+      margin-left: 0;
+    }
+  }
 }
+
 body > main > section.fitting__section > div > div.fitting__video-slider > h3:before {
   content: '/';
   font-weight: 500;
@@ -75,48 +87,6 @@ body > main > section.fitting__section > div > div.fitting__video-slider > h3:be
   right: 6%;
   @media screen and (min-width: $mobile-first) and (max-width: $tablet-general) {
     display: none;
-  }
-}
-
-.slider .slick-arrow.slick-next {
-  right: 10px;
-  top: -40px;
-  background-image: url('../images/arrow-right.svg');
-  background-repeat: no-repeat;
-  width: 40px;
-  height: 15px;
-  border: none;
-  background-color: #f9f9f9;
-}
-.slider .slick-dots {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-}
-.slider .slick-dots li {
-  list-style: none;
-}
-.slider .slick-dots button {
-  font-size: 0;
-  width: 5px;
-  height: 5px;
-  background-color: $in-instagram-slider-dots;
-  border-radius: 70%;
-  border: none;
-}
-.slider .slick-dots li.slick-active button {
-  width: 26px;
-  border-radius: 20%;
-  @media screen and (min-width: $desktop-up) {
-    width: 66px;
-  }
-}
-.slider__item {
-  position: relative;
-
-  @media screen and (min-width: $mobile-first) {
-    margin-left: 0;
   }
 }
 
@@ -140,7 +110,6 @@ body > main > section.fitting__section > div > div.fitting__video-slider > h3:be
     right: 18px;
   }
 }
-
 
 .video-fitting-fullscreen {
   display: none;
