+++ src/sass/utils/_variabless.scss
@@ -52,7 +52,7 @@ $accent-colorpicker-grey: #e2dfdf;
 $third-accent-color: #c0a78f;
 $accent-background-selector-color: #f8efe7;
 $button-hover-background-color: #000000;
-$accent-footer-color:#d7b494;
+$accent-footer-color: #d7b494;
 
 //background
 $primary-background-color: #ffffff;
@@ -70,10 +70,11 @@ $backdrop-background: rgba(0, 0, 0, 0.6);
 $btn-background-color: #efefef;
 $message-amount-thumb-color: #e20613;
 $favorites-after-color: #eaeaea;
+$predownload-background: #d9dbe1;
 
 //svg-fill
 $primary-svg-color: #000000;
-$secondary-svg-color:#2f2f2f;
+$secondary-svg-color: #2f2f2f;
 
 //logo
 $primary-logo-color: #000000;
