+++ src/views/partials/home/inInstagram.hbs
@@ -18,12 +18,14 @@
     {{image-mobile-1x}} 1x,
     {{image-mobile-2x}} 2x'
                   media='(max-width: 1377px)'
+                  width="300" height="478"
                 />
                 <source
                   srcset='
   {{image-desktop-1x}} 1x,
   {{image-desktop-2x}} 2x'
                   media='(min-width: 1378px)'
+                  width="260" height="438"
                 />
                 <img
                   class='in-instagram__title-image'
