+++ src/views/partials/home/ourShowRoom.hbs
@@ -9,15 +9,17 @@
           <picture>
             <source
               srcset='
-    {{image-mobile-1x}} 1x,
-    {{image-mobile-2x}} 2x'
+              {{image-mobile-1x}} 1x,
+              {{image-mobile-2x}} 2x'
               media='(max-width: 1377px)'
+              width="336" height="460"
             />
             <source
               srcset='
-  {{image-desktop-1x}} 1x,
-  {{image-desktop-2x}} 2x'
+              {{image-desktop-1x}} 1x,
+              {{image-desktop-2x}} 2x'
               media='(min-width: 1378px)'
+              width="426" height="502"
             />
             <img
               class='showroom__title-image animate__animated animate__pulse animate__repeat-1'
@@ -25,8 +27,6 @@
               alt='{{name-title}}'
               id='{{id-title-image}}'
               loading='lazy'
-              width='374px'
-              height='512px'
               ;
             />
           </picture>
