+++ src/views/partials/home/starClients.hbs
@@ -11,12 +11,14 @@
     {{image-mobile-1x}} 1x,
     {{image-mobile-2x}} 2x'
               media='(max-width: 1377px)'
+              width="290" height="420"
             />
             <source
               srcset='
   {{image-desktop-1x}} 1x,
   {{image-desktop-2x}} 2x'
               media='(min-width: 1378px)'
+              width="256" height="370"
             />
             <img src='#' alt='{{name}}' id='{{id}}' loading='lazy' width='256px' height='370px' />
           </picture>
