+++ src/views/partials/product/infoAboutProduct.hbs
@@ -8,7 +8,7 @@
       <article class="product__content">
         <div class="product__main-info">
           <div class="product__name-wrapper">
-            <h2 class="product__name">{{parsedProductInfoData.productName}}</h2>
+            <h2 class="product__name">{{parsedProductInfoData.name}}</h2>
             {{> ../../components/buttonAddLike}}
           </div>
             <p class="product__price">{{parsedProductInfoData.productPrice}} <span class="product__currency">&#8372;</span></p>
